Add createAccount endpoint

+++ fintech-bank/src/main/java/com/javabootcamp/fintechbank/accounts/CreateAccountRequest.java
@@ -5,13 +5,14 @@ import jakarta.validation.constraints.Size;
 
 public record CreateAccountRequest(
 
-        @NotNull String type,
+        @NotNull
+        String type,
         @NotNull
         @Size(min = 3, max = 50, message = "name must be between 3 and 50 characters")
         String name,
 
         @NotNull
-        Integer balance
+        Double balance
 
         ){
 }
